+++ example/lib/pages/alert_page.dart
@@ -1,6 +1,7 @@
 import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:example/router.dart';
 import 'package:example/util/logger.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 class AlertPage extends StatelessWidget {
@@ -171,6 +172,35 @@ class AlertPage extends StatelessWidget {
                   ),
                 ],
                 shrinkWrap: false,
+              );
+              logger.info(result);
+            },
+          ),
+          ListTile(
+            title: const Text('OK/Cancel Dialog RichText'),
+            onTap: () async {
+              final result = await showOkCancelAlertDialog(
+                context: context,
+                title: 'Title',
+                richMessage: RichText(
+                  text: TextSpan(
+                      style: const TextStyle(
+                        color: Colors.black,
+                        fontSize: 14,
+                      ),
+                      children: <TextSpan>[
+                        const TextSpan(text: '登陆即视为同意'),
+                        TextSpan(
+                          text: '《xxx服务协议》',
+                          style: const TextStyle(color: Colors.red),
+                          recognizer: TapGestureRecognizer()
+                            ..onTap = () {
+                              // Add your own processing
+                            },
+                        ),
+                      ]),
+                ),
+                alertStyle: AdaptiveStyle.cupertino,
               );
               logger.info(result);
             },